+++ examples/create-proposal/index.js
@@ -1,10 +1,10 @@
 require('dotenv').config();
 
-const { Platform } = require('@openzeppelin/defender-sdk');
+const { Defender } = require('@openzeppelin/defender-sdk');
 
 async function main() {
   const creds = { apiKey: process.env.API_KEY, apiSecret: process.env.API_SECRET };
-  const client = new Platform(creds);
+  const client = new Defender(creds);
 
   const proposal = await client.proposal.create({
     proposal: {