+++ examples/simulate-proposal/index.js
@@ -1,13 +1,13 @@
 require('dotenv').config();
 
-const { Platform } = require('@openzeppelin/defender-sdk');
+const { Defender } = require('@openzeppelin/defender-sdk');
 const { utils } = require('ethers');
 
 const contractABI = require('./abi/demoflash.json');
 
 async function main() {
   const creds = { apiKey: process.env.API_KEY, apiSecret: process.env.API_SECRET };
-  const client = new Platform(creds);
+  const client = new Defender(creds);
 
   // Create a new proposal
   const proposal = await client.proposal.create({