+++ packages/action/src/api.ts
@@ -12,16 +12,16 @@ type SourceFiles = {
 
 export class ActionClient extends BaseApiClient {
   protected getPoolId(): string {
-    return process.env.PLATFORM_POOL_ID || 'us-west-2_94f3puJWv';
+    return process.env.DEFENDER_POOL_ID || 'us-west-2_94f3puJWv';
   }
 
   protected getPoolClientId(): string {
-    return process.env.PLATFORM_POOL_CLIENT_ID || '40e58hbc7pktmnp9i26hh5nsav';
+    return process.env.DEFENDER_POOL_CLIENT_ID || '40e58hbc7pktmnp9i26hh5nsav';
   }
 
   protected getApiUrl(): string {
-    // TODO: update to platform-api.* url (and /action) when available
-    return process.env.PLATFORM_API_URL || 'https://defender-api.openzeppelin.com/autotask/';
+    // TODO: update to /action when available
+    return process.env.DEFENDER_API_URL || 'https://defender-api.openzeppelin.com/autotask/';
   }
 
   public async list(): Promise<ActionListResponse> {