+++ packages/base/src/api/__mocks__/axios-error.ts
@@ -13,7 +13,7 @@ export const mockAxiosError: AxiosError = {
       'Content-Type': 'application/json',
       'User-Agent': 'axios/0.21.4',
     } as unknown) as AxiosRequestHeaders,
-    baseURL: 'https://platform-api.openzeppelin.com/',
+    baseURL: 'https://defender-api.openzeppelin.com/',
     timeout: 0,
     xsrfCookieName: 'XSRF-TOKEN',
     xsrfHeaderName: 'X-XSRF-TOKEN',
@@ -54,7 +54,7 @@ export const mockAxiosError: AxiosError = {
       'Authorization: Bearer WRONG\r\n' +
       'Content-Type: application/json\r\n' +
       'User-Agent: axios/0.21.4\r\n' +
-      'Host: platform-api.openzeppelin.com\r\n' +
+      'Host: defender-api.openzeppelin.com\r\n' +
       'Connection: close\r\n' +
       '\r\n',
     _keepAliveTimeout: 0,
@@ -97,7 +97,7 @@ export const mockAxiosError: AxiosError = {
       statusMessage: 'Unauthorized',
       _consuming: false,
       _dumped: false,
-      responseUrl: 'https://platform-api.openzeppelin.com/relayer',
+      responseUrl: 'https://defender-api.openzeppelin.com/relayer',
       redirects: [],
     },
     aborted: false,
@@ -106,7 +106,7 @@ export const mockAxiosError: AxiosError = {
     parser: null,
     maxHeadersCount: null,
     reusedSocket: false,
-    host: 'platform-api.openzeppelin.com',
+    host: 'defender-api.openzeppelin.com',
     protocol: 'https:',
   },
   response: {
@@ -133,7 +133,7 @@ export const mockAxiosError: AxiosError = {
     config: {
       url: '/relayer',
       method: 'get',
-      baseURL: 'https://platform-api.openzeppelin.com/',
+      baseURL: 'https://defender-api.openzeppelin.com/',
       transformRequest: [Array],
       transformResponse: [Array],
       timeout: 0,
@@ -177,7 +177,7 @@ export const mockAxiosError: AxiosError = {
         'Authorization: Bearer WRONG\r\n' +
         'Content-Type: application/json\r\n' +
         'User-Agent: axios/0.21.4\r\n' +
-        'Host: platform-api.openzeppelin.com\r\n' +
+        'Host: defender-api.openzeppelin.com\r\n' +
         'Connection: close\r\n' +
         '\r\n',
       _keepAliveTimeout: 0,
@@ -192,7 +192,7 @@ export const mockAxiosError: AxiosError = {
       parser: null,
       maxHeadersCount: null,
       reusedSocket: false,
-      host: 'platform-api.openzeppelin.com',
+      host: 'defender-api.openzeppelin.com',
       protocol: 'https:',
     },
     data: { message: 'Unauthorized' },