+++ packages/base/src/api/api.error.test.ts
@@ -1,6 +1,6 @@
 import { pick } from 'lodash';
-import { rejectWithPlatformApiError } from './api';
-import { PlatformApiResponseError } from './api-error';
+import { rejectWithDefenderApiError } from './api';
+import { DefenderApiResponseError } from './api-error';
 import { mockAxiosError } from './__mocks__/axios-error';
 
 const expectedRejectObjectStructure = {
@@ -14,12 +14,12 @@ const expectedRejectObjectStructure = {
 };
 
 // prettier-ignore
-describe('PlatformApiError', () => {
-  test('request rejection reject with a PlatformApiResponseError that include message, request.path, request.method, response.status, response.statusText, response.data', async () => {
+describe('DefenderApiError', () => {
+  test('request rejection reject with a DefenderApiResponseError that include message, request.path, request.method, response.status, response.statusText, response.data', async () => {
     try {
-      await rejectWithPlatformApiError(mockAxiosError);
+      await rejectWithDefenderApiError(mockAxiosError);
     } catch (error: any) {
-      expect(error instanceof PlatformApiResponseError).toBe(true);
+      expect(error instanceof DefenderApiResponseError).toBe(true);
 
       expect(error.message).toStrictEqual(expectedRejectObjectStructure.message);
       expect(error.response).toStrictEqual(expectedRejectObjectStructure.response);