+++ packages/base/src/index.ts
@@ -6,5 +6,4 @@ export * from './utils/network';
 // eslint-disable-next-line @typescript-eslint/no-var-requires
 export const VERSION = require('../package.json').version;
 
-// TODO: Update to platform.* when available
-export const PLATFORM_APP_URL = 'https://defender.openzeppelin.com';
+export const DEFENDER_APP_URL = 'https://defender.openzeppelin.com';