+++ packages/deploy/src/api/index.ts
@@ -23,15 +23,15 @@ const DEPLOYMENTS_CONFIG_PATH = '/deployment-config';
 
 export class DeployClient extends BaseApiClient {
   protected getPoolId(): string {
-    return process.env.PLATFORM_POOL_ID || 'us-west-2_94f3puJWv';
+    return process.env.DEFENDER_POOL_ID || 'us-west-2_94f3puJWv';
   }
 
   protected getPoolClientId(): string {
-    return process.env.PLATFORM_POOL_CLIENT_ID || '40e58hbc7pktmnp9i26hh5nsav';
+    return process.env.DEFENDER_POOL_CLIENT_ID || '40e58hbc7pktmnp9i26hh5nsav';
   }
 
   protected getApiUrl(): string {
-    return process.env.PLATFORM_API_URL || 'https://defender-api.openzeppelin.com/deployment/';
+    return process.env.DEFENDER_API_URL || 'https://defender-api.openzeppelin.com/deployment/';
   }
 
   public async deployContract(payload: DeployContractRequest): Promise<DeploymentResponse> {