+++ packages/proposal/src/api/index.ts
@@ -21,15 +21,15 @@ type SimulateProposalParams = {
 
 export class ProposalClient extends BaseApiClient {
   protected getPoolId(): string {
-    return process.env.PLATFORM_POOL_ID || 'us-west-2_94f3puJWv';
+    return process.env.DEFENDER_POOL_ID || 'us-west-2_94f3puJWv';
   }
 
   protected getPoolClientId(): string {
-    return process.env.PLATFORM_POOL_CLIENT_ID || '40e58hbc7pktmnp9i26hh5nsav';
+    return process.env.DEFENDER_POOL_CLIENT_ID || '40e58hbc7pktmnp9i26hh5nsav';
   }
 
   protected getApiUrl(): string {
-    return process.env.PLATFORM_API_URL || 'https://defender-api.openzeppelin.com/proposal/';
+    return process.env.DEFENDER_API_URL || 'https://defender-api.openzeppelin.com/proposal/';
   }
 
   public async addContract(contract: Contract): Promise<Contract> {