+++ packages/proposal/src/api/utils.ts
@@ -1,7 +1,7 @@
 import { ExternalApiProposalResponse } from '../models/response';
 
-import { PLATFORM_APP_URL } from '@openzeppelin/defender-sdk-base-client';
+import { DEFENDER_APP_URL } from '@openzeppelin/defender-sdk-base-client';
 
 export function getProposalUrl(proposal: Pick<ExternalApiProposalResponse, 'contractId' | 'proposalId'>): string {
-  return `${PLATFORM_APP_URL}/#/admin/contracts/${proposal.contractId}/proposals/${proposal.proposalId}`;
+  return `${DEFENDER_APP_URL}/#/admin/contracts/${proposal.contractId}/proposals/${proposal.proposalId}`;
 }