+++ packages/relay-signer/src/models/relayer.ts
@@ -3,7 +3,7 @@ import { Network } from '@openzeppelin/defender-sdk-base-client';
 import { ListTransactionsRequest, RelayerTransaction, RelayerTransactionPayload } from './transactions';
 import { JsonRpcResponse, SignMessagePayload, SignTypedDataPayload, SignedMessagePayload } from './rpc';
 
-// TODO platform Address model for this
+// TODO Defender Address model for this
 export type Address = string;
 
 export type BigUInt = string | number;